[Tickets:21989] Make sure the folder exists when trying to delete the cached images.
git-svn-id: https://teamsupport.svn.beanstalkapp.com/main/trunk@3422 02885c34-e171-44ab-8435-42fb7a3c3aaf

+++ TeamSupport.ServiceLibrary/CustomerInsightsProcessor.cs
@@ -317,12 +317,17 @@ namespace TeamSupport.ServiceLibrary
             try
             {
               cachePath = System.IO.Path.Combine(AttachmentPath.GetImageCachePath(LoginUser), "CompanyLogo\\" + currentCompanyInfo.ParentID.ToString());
-              pattern = currentCompanyInfo.OrganizationID.ToString() + "-*.*";
-              string[] files = System.IO.Directory.GetFiles(cachePath, pattern, System.IO.SearchOption.TopDirectoryOnly);
 
-              foreach (String file in files)
+              if (System.IO.Directory.Exists(cachePath))
               {
-                System.IO.File.Delete(file);
+                pattern = currentCompanyInfo.OrganizationID.ToString() + "-*.*";
+                string[] files = System.IO.Directory.GetFiles(cachePath, pattern, System.IO.SearchOption.TopDirectoryOnly);
+
+                foreach (String file in files)
+                {
+                  System.IO.File.Delete(file);
+                  Logs.WriteEvent(string.Format("Cached file {0} deleted.", file));
+                }
               }
             }
             catch (Exception ex)
